Proper comments and location added
